Refactor transaction amount handling to use MoneyAmount structure
- Updated transaction-related tests to utilize MoneyAmount for amount representation.
- Changed amount fields in AddTransactionRequest, UpdateTransactionRequest, and other entities to use amountMinor and amountScale.
- Introduced MoneyAmountJsonConverter for JSON serialization of MoneyAmount.
- Adjusted balance representation in AccountEntity to use balanceMinor and currencyScale.
- Ensured consistency across upcoming payments and payment reminders to use MoneyAmount for amounts.

+++ lib/features/accounts/presentation/account_details_screen.dart
@@ -343,7 +343,9 @@ class _AccountSummaryCard extends StatelessWidget {
                 Expanded(
                   child: _SummaryValue(
                     label: strings.accountDetailsIncomeLabel,
-                    value: currencyFormat.format(summary.totalIncome),
+                    value: currencyFormat.format(
+                      summary.totalIncome.toDouble(),
+                    ),
                     valueColor: theme.colorScheme.primary,
                   ),
                 ),
@@ -351,7 +353,9 @@ class _AccountSummaryCard extends StatelessWidget {
                 Expanded(
                   child: _SummaryValue(
                     label: strings.accountDetailsExpenseLabel,
-                    value: currencyFormat.format(summary.totalExpense),
+                    value: currencyFormat.format(
+                      summary.totalExpense.toDouble(),
+                    ),
                     valueColor: theme.colorScheme.error,
                   ),
                 ),
